+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 - Clear exchange/symbol info cache in the Redis periodically - [#284](https://github.com/chrisleekr/binance-trading-bot/issues/284)
+- Added minimum required order amount - [#84](https://github.com/chrisleekr/binance-trading-bot/issues/84)
 
 ## [0.0.78] - 2021-09-05